Porting.

+++ Box2D.Net/Common/b2Settings.cs
@@ -39,7 +39,7 @@ namespace Box2D {
 
 		/// A small angle used as a collision and constraint tolerance. Usually it is
 		/// chosen to be numerically significant, but visually insignificant.
-		public const float b2_angularSlop		=	(2.0f / 180.0f * Math.PI);
+		public const float b2_angularSlop		=	(float)(2.0f / 180.0f * Math.PI);
 
 		/// The radius of the polygon/edge shape skin. This should not be modified. Making
 		/// this smaller means polygons will have an insufficient buffer for continuous collision.
